XMLHttpRequest/tests/submissions/Microsoft/abort-event-order.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The abort() method: abort and loadend events</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        var test = async_test();

        test.step(function()
        {
            var xhr = new XMLHttpRequest();
            var initialized = false;
            var expect = [4, "upload.abort", "upload.loadend", "abort", "loadend"];
            var actual = [];

            xhr.onreadystatechange = function()
            {
                test.step(function()
                {
                    var readyState = xhr.readyState;
                    
                    if (readyState == 1)
                    {
                        if (initialized) 
                        {
                            xhr.abort();
                        }
                        initialized = true;
                    }
                    else if (readyState == 4) 
                    {
                        actual.push(readyState);
                    }
                });
            };
            
            xhr.onloadend          = function(e){ actual.push(e.type); VerifyResult()};     
            xhr.onabort            = function(e){ actual.push(e.type); };

            xhr.upload.onloadend   = function(e){ actual.push("upload." + e.type);};
            xhr.upload.onabort     = function(e){ actual.push("upload." + e.type); };

            function VerifyResult()
            {
                test.step(function()
                {
                    assert_array_equals(actual, expect);
                    test.done();
                });
            };

            xhr.open("POST", "./support/echo.php", true);
            xhr.send("Test Message");
        });
    </script>
</body>
</html>
